refactor: convert to webp
- set all attached images to webp
- set in attaching files
