+++ src/data/2025/jobboard.js
@@ -1,10 +1,10 @@
 // Job Board data for displaying companies and their job opportunities
-import ccsLogo from '@/data/2025/assets/images/job-board/ccs-logo.png'
-import dteLogo from '@/data/2025/assets/images/job-board/dte-logo.png'
-import grandCircusLogo from '@/data/2025/assets/images/job-board/grand-circus-logo.png'
-import compassLogo from '@/data/2025/assets/images/job-board/compass-logo.png'
-import googleLogo from '@/data/2025/assets/images/job-board/google-logo.png'
-import ibmLogo from '@/data/2025/assets/images/job-board/ibm-logo.png'
+import ccsLogo from '@/data/2025/assets/images/job-board/ccs-logo.webp'
+import dteLogo from '@/data/2025/assets/images/job-board/dte-logo.webp'
+import grandCircusLogo from '@/data/2025/assets/images/job-board/grand-circus-logo.webp'
+import compassLogo from '@/data/2025/assets/images/job-board/compass-logo.webp'
+import googleLogo from '@/data/2025/assets/images/job-board/google-logo.webp'
+import ibmLogo from '@/data/2025/assets/images/job-board/ibm-logo.webp'
 
 export const jobBoardData = [
   {
