+++ src/data/2025/partners.js
@@ -2,12 +2,12 @@ import CompassDetroit from '@/assets/images/sponsors/Compass_Detroit_logo.webp'
 import Google from '@/assets/images/sponsors/Google_logo.webp'
 import SHPEDetroit from '@/assets/images/organizations/org-shpe-logo.webp'
 import WomenTechmakers from '@/assets/images/organizations/org-wtm-logo.webp'
-import GDGDetroit from '@/assets/images/organizations/org-gdg-detroit.png'
-import GDGAnnArbor from '@/assets/images/organizations/org-gdg-a2.png'
-import GDGToledo from '@/assets/images/organizations/org-gdg-toledo.png'
-import GDGWindsor from '@/assets/images/organizations/org-gdg-windsor.png'
-import GDGUofMDearborn from '@/assets/images/organizations/org-gdg-umd.png'
-import GDGWSU from '@/assets/images/organizations/org-gdg-wsu.png'
+import GDGDetroit from '@/assets/images/organizations/org-gdg-detroit.webp'
+import GDGAnnArbor from '@/assets/images/organizations/org-gdg-a2.webp'
+import GDGToledo from '@/assets/images/organizations/org-gdg-toledo.webp'
+import GDGWindsor from '@/assets/images/organizations/org-gdg-windsor.webp'
+import GDGUofMDearborn from '@/assets/images/organizations/org-gdg-umd.webp'
+import GDGWSU from '@/assets/images/organizations/org-gdg-wsu.webp'
 import NSBEDetroit from '@/assets/images/organizations/org-nsbe-logo.webp'
 import WSU from '@/assets/images/sponsors/spo-wsu-logo.webp'
 import IBM from '@/assets/images/sponsors/spo-ibm-logo.webp'
